[bug][Android] More NPE protection for casting operations
Sometimes when the chromecast device has not yet received media info from the stream it is returned as null from the client. This can lead to NPEs during multiple operations (usually when dealing with MediaStatus). The remote client itself can sometimes be returned as null from the session, usually when the device has disconnected but the session has not been marked as ended by the SDK yet. In both cases we add extra NPE protection that will result in no-ops if the required objects are null.