Update package.json to show contribs over author
Hi! I'm the original author of this library, and even if I haven't been involved in the latest years, my work was the foundation of this project

Author in the package.json is a misleading property, and I value the work of all the people involved in this project

I returned the contributors field, removed the author field because it only allows one name

Contributors field honors that more than one person works in this project

(More reasons to use contributors over author field here: https://github.com/schibsted/eslint-config-schibsted/issues/41)

Again, thank you all for your great work. Happy to see that this library, which was my side project a couple of years ago, has evolved so strongly over the years :)